Items & Accounts initialized. Authentication updated

+++ testProject/fictionBook.h
@@ -6,7 +6,7 @@
 
 class FictionBook : public Book{
     public:
-        FictionBook(const std::string& name, const std::string& author, int year, int isbn = 0);
+        FictionBook(const std::string& name, const std::string& author, int year, long isbn = 0);
         virtual ~FictionBook();
         virtual void display() const;
     private:
