+++ testProject/movie.h
@@ -5,7 +5,7 @@
 
 class Movie : public Item{
     public:
-        Movie(const std::string& name, const std::string& author, int year, const std::string genre, int rating = -1, int isbn = 0);
+        Movie(const std::string& name, const std::string& author, int year, const std::string genre, int rating = -1, long isbn = 0);
         virtual ~Movie();
         virtual void display() const;
     private:
