+++ testProject/nonFictionBook.h
@@ -5,7 +5,7 @@
 
 class NonFictionBook : public Book{
     public:
-        NonFictionBook(const std::string& name, const std::string& author, const std::string& dewyD, int year, int isbn = 0);
+        NonFictionBook(const std::string& name, const std::string& author, const std::string& dewyD, int year, long isbn = 0);
         virtual ~NonFictionBook();
         virtual void display() const;
     private:
