+++ testProject/videoGame.h
@@ -5,7 +5,7 @@
 
 class VideoGame : public Item{
     public:
-        VideoGame(const std::string& name, const std::string& author, int year, const std::string genre, int rating = -1, int isbn = 0);
+        VideoGame(const std::string& name, const std::string& author, int year, const std::string genre, int rating = -1, long isbn = 0);
         virtual ~VideoGame();
         virtual void display() const;
     private:
